% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETScan.plot.R
\name{PETScan.plot}
\alias{PETScan.plot}
\title{Visualization of significant gene-peak pairs of score-based genome-wide association analysis}
\usage{
PETScan.plot(
  geneID,
  peakID,
  scores,
  scuts,
  cols,
  eChr,
  ePos,
  pChr,
  pPos,
  chroms,
  plot.peak = TRUE,
  plot.gene = TRUE,
  xlab = "Peak Location",
  ylab = "Gene Location"
)
}
\arguments{
\item{geneID}{A vector of gene IDs from significant gene-peak pairs.
If geneID=k, its chromosome and position will be extracted from `eChr[k]` and `ePos[k]`, respectively.}

\item{peakID}{A vector of peak IDs from significant gene-peak pairs.
If peakID=k, its chromosome and position will be extracted from `pChr[k]` and `pPos[k]`, respectively.}

\item{scores}{A vector of p-values of each significant gene-peak pair.}

\item{scuts}{A vector indicating cutoffs of different significance levels.}

\item{cols}{A vector indicating colors for different significance levels.}

\item{eChr}{A vector specifying the chromosomes of genes, ordered by gene IDs.}

\item{ePos}{A vector specifying the genomic positions of the genes 
on their respective chromosomes, ordered by gene IDs.}

\item{pChr}{A vector specifying the chromosomes of the peaks, ordered by peak IDs.}

\item{pPos}{A vector specifying the genomic positions of the peaks 
on their respective chromosomes, ordered by peak IDs.}

\item{chroms}{A vector indicating the all the distinct chromosomes visualized.}

\item{plot.peak}{A logical indicating whether to count the number of peaks linked to each gene.}

\item{plot.gene}{A logical indicating whether to count the number of genes linked to each peak.}
}
\description{
PETScan.plot() shows significant gene-peak pairs of score-based genome-wide 
association analysis using negative binomial models, 
with peak locations on the X-axis and gene locations on the Y-axis. 
Different levels of significance are indicated using different colors. 
The middle panel illustrates the number of genes associated with each peak, 
while the bottom panel displays the number of peaks linked to each gene.
}
\examples{

RNA = ENCODE_RNA
ATAC = ENCODE_ATAC
cvrt = ENCODE_cvrt[,2:4]
nb1 = PETScan(Ymat=RNA, X=cvrt, A=ATAC, perm=2, no_cores=8, seed=1) 

library(dplyr)
# Apply genomic control methods
nrow1 = nrow(ATAC)
score1 = nb1[1:nrow1, ]
gc1 = nb1[nrow1+1, ] / 0.456
gc1 = ifelse(gc1>1, gc1, 1)
score2 = t(t(score1) / gc1) \%>\% data.frame()
score2$ATAC = rownames(score2)

RNApos = ENCODE_RNApos
ATACpos = ENCODE_ATACpos
score3 = reshape2::melt(score2, id.vars="ATAC") \%>\%
  mutate(pvalue = 1 - pchisq(value, 1)) \%>\%
  mutate(padj = p.adjust(pvalue, method="BH")) \%>\%
  left_join(RNApos, by=c("variable"="gene")) \%>\%
  left_join(ATACpos, by=c("ATAC"="peak"))

PETScan.plot(score3$geneid, score3$peakid, score3$padj, 
          scuts = c(0.2, 0.1, 0.05, 0.01), 
          cols = c("yellow", "green", "blue", "red"),
          eChr = RNApos$g.chr, ePos = RNApos$g.start,
          pChr = ATACpos$p.chr, pPos = ATACpos$p.start, 
          chroms = c(1:22, "X"))

}
\references{
Sun, Wei. (2009). eQTL analysis by Linear Model. http://www.bios.unc.edu/~weisun/software/eMap.pdf
}
