% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETScan.R
\name{PETScan}
\alias{PETScan}
\title{Perform score-based genome-wide association analysis using negative binomial models}
\usage{
PETScan(Ymat, X = NULL, A, perm, no_cores = 64, seed = 1)
}
\arguments{
\item{Ymat}{A matrix of gene expression for RNA-Seq genes.
Rows and columns represent RNA-Seq gene names and samples, respectively.}

\item{X}{A matrix of covariates.}

\item{A}{A matrix of chromatin accessibility for ATAC-Seq peaks.
Rows and columns represent ATAC-Seq peak positions and samples, respectively.}

\item{perm}{An integer indicating the number of permutations for genomic control.}

\item{no_cores}{An integer indicating the number of cores for parallel calculations.}

\item{seed}{An integer specifying the seed for random number generation.}
}
\value{
A matrix of score test statistics across all RNA-Seq genes and ATAC-Seq peaks.
Rows and columns represent ATAC-Seq peak positions and RNA-Seq gene names, respectively.
The last row contains the median of permuted score test statistics for each gene.
}
\description{
PETScan() performs score-based genome-wide association analysis
across all RNA-Seq genes and ATAC-Seq peaks using negative binomial models,
and returns a matrix where the (i,j) element of the matrix represents
the score test statistic for the i-th ATAC-Seq peak and the j-th RNA-Seq gene.
}
\examples{
RNA = ENCODE_RNA
ATAC = ENCODE_ATAC
cvrt = ENCODE_cvrt[,2:4]
nb1 = PETScan(Ymat=RNA, X=cvrt, A=ATAC, perm=2, no_cores=8, seed=1)

}
