% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETScan_Mixed.R
\name{PETScan_Mixed}
\alias{PETScan_Mixed}
\title{Perform score-based genome-wide association analysis using negative binomial mixed models}
\usage{
PETScan_Mixed(Ymat, X = NULL, R, A, perm, no_cores = 64, seed = 1, r0 = 1)
}
\arguments{
\item{Ymat}{A matrix of gene expression for RNA-Seq genes.
Rows and columns represent RNA-Seq gene names and samples, respectively.}

\item{X}{A matrix of covariates.}

\item{R}{A vector indicating the group of samples.}

\item{A}{A matrix of chromatin accessibility for ATAC-Seq peaks.
Rows and columns represent ATAC-Seq peak positions and samples, respectively.}

\item{perm}{An integer indicating the number of permutations for genomic control.}

\item{no_cores}{An integer indicating the number of cores for parallel calculations.}

\item{seed}{An integer specifying the seed for random number generation.}

\item{r0}{A decimal specifying the threshold for reliability check.}
}
\value{
A matrix of score test statistics across all RNA-Seq genes and ATAC-Seq peaks.
Rows and columns represent ATAC-Seq peak positions and RNA-Seq gene names, respectively.
The last row contains the median of permuted score test statistics for each gene.
}
\description{
PETScan_Mixed() performs score-based genome-wide association analysis
across all RNA-Seq genes and ATAC-Seq peaks using negative binomial mixed models,
and returns a matrix where the (i,j) element of the matrix represents
the score test statistic for the i-th ATAC-Seq peak and the j-th RNA-Seq gene.
}
\examples{
RNA = GSE118165_RNA
ATAC = GSE118165_ATAC
cvrt = GSE118165_cvrt
nbm1 = PETScan_Mixed(Ymat=RNA, X=cvrt[,1:2], R=cvrt[,5], A=ATAC, perm=2, no_cores=8, seed=1)

}
