% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETScan.R
\name{scorep.nb}
\alias{scorep.nb}
\title{Compute score test statistics for a given gene across peaks using negative binomial models}
\usage{
scorep.nb(Y, X, A, Ap)
}
\arguments{
\item{Y}{A vector of gene expression for a given RNA-Seq gene.}

\item{X}{A matrix of covariates.}

\item{A}{A matrix of chromatin accessibility for ATAC-Seq peaks.
Rows and columns represent ATAC peak positions and samples, respectively.}

\item{Ap}{A matrix of permuted chromatin accessibility for ATAC-Seq peaks.}
}
\value{
A vector of score test statistics for a given gene across all ATAC-Seq peaks
and the median of permuted score test statistics.
}
\description{
scorep.nb() computes score test statistics for a given gene
across all ATAC-Seq peaks using negative binomial models,
and returns a vector where the i-th element of the vector represents
the score test statistic for the i-th ATAC-Seq peak and the given gene.
}
