% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PETScan_Mixed.R
\name{scorep.nbm}
\alias{scorep.nbm}
\title{Compute score test statistics for a given gene across peaks using negative binomial mixed models}
\usage{
scorep.nbm(Y, X, R, A, Ap, r0)
}
\arguments{
\item{Y}{A vector of gene expression for a given RNA-Seq gene.}

\item{X}{A matrix of covariates.}

\item{R}{A vector indicating the group of samples.}

\item{A}{A matrix of chromatin accessibility for ATAC-Seq peaks.
Rows and columns represent ATAC-Seq peak positions and samples, respectively.}

\item{Ap}{A matrix of permuted chromatin accessibility for ATAC-Seq peaks.}

\item{r0}{A decimal specifying the filtering threshold.}
}
\value{
A vector of score test statistics for a given gene across all ATAC-Seq peaks
and the median of permuted score test statistics.
}
\description{
scorep.nbm() computes score test statistics for a given gene
across all ATAC-Seq peaks using negative binomial mixed models,
and return a vector where the i-th element of the vector represents
the score test statistic for the i-th ATAC-Seq peak and the given gene.
}
